refacto utils
